+++ src/main/java/org/relunluck/engine/scene/Projection.java
@@ -1,25 +1,25 @@
 package org.relunluck.engine.scene;
 
-import org.joml.Matrix4f;
+import org.joml.Matrix4d;
 
 public class Projection {
 
-    private static final float FOV = (float) Math.toRadians(60.0f);
-    private static final float Z_FAR = 1000.f;
-    private static final float Z_NEAR = 0.01f;
+    private static final double FOV = Math.toRadians(60.0d);
+    private static final double Z_FAR = 1000.d;
+    private static final double Z_NEAR = 0.01d;
 
-    private Matrix4f projMatrix;
+    private Matrix4d projMatrix;
 
     public Projection(int width, int height) {
-        projMatrix = new Matrix4f();
+        projMatrix = new Matrix4d();
         updateProjMatrix(width, height);
     }
 
-    public Matrix4f getProjMatrix() {
+    public Matrix4d getProjMatrix() {
         return projMatrix;
     }
 
     public void updateProjMatrix(int width, int height) {
-        projMatrix.setPerspective(FOV, (float) width / height, Z_NEAR, Z_FAR);
+        projMatrix.setPerspective(FOV, (double) width / height, Z_NEAR, Z_FAR);
     }
 }
